// clients/web/application/src/app/core/models/api.d.ts
/* tslint:disable */
// Generated using typescript-generator version 2.12.476 on 2019-04-11 11:18:25.

export interface AccountDTO {
    idToken: string;
    firstName: string;
    lastName: string;
    email: string;
    pictureUrl: string;
    newAccount: boolean;
}

export interface ContactDTO {
    email: string;
    firstName: string;
    lastName: string;
    notes: string;
    contactRecurrenceType: string;
    contactRecurrenceValue: number;
    contactDueDate: Date;
}